Finish success action by re-opening modal
