+++ front/app/containers/Authentication/SuccessActions/actions/attendEvent.ts
@@ -3,6 +3,7 @@ import { IUserData } from 'api/users/types';
 
 import { addEventAttendance } from 'api/event_attendance/useAddEventAttendance';
 import { addFollower } from 'api/follow_unfollow/useAddFollower';
+import eventEmitter from 'utils/eventEmitter';
 
 export interface AttendEventParams {
   event: IEventData;
@@ -10,10 +11,14 @@ export interface AttendEventParams {
 
 export const attendEvent = ({ event }: AttendEventParams) => {
   return async (authUser: IUserData) => {
-    addEventAttendance({ eventId: event.id, attendeeId: authUser.id });
-    addFollower({
-      followableId: event.relationships.project.data.id,
-      followableType: 'projects',
-    });
+    await Promise.all([
+      addEventAttendance({ eventId: event.id, attendeeId: authUser.id }),
+      addFollower({
+        followableId: event.relationships.project.data.id,
+        followableType: 'projects',
+      }),
+    ]);
+
+    eventEmitter.emit('eventAttendance');
   };
 };
